Ngoc Quyen, Quynh Tram, Mac Nguyen chinh giao dien
